+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog of OcLaunch
 
+## v0.2.1
+ +
+
 ## v0.2.0
  + Add new command line option.
    + “-r” can now take a number to start from.